Tests/LibWeb/Text/expected/css/inline-block-aspect-ratio-definite-height.txt

auto: 40, 20
percentage: 40, 20
intrinsic: 40, 20
button: 40, 15
